+++ packages/library/src/__tests__/rpc-websocket-autopinger-test.ts
@@ -1,4 +1,4 @@
-import type { IRpcWebSocketTransport } from '@solana/rpc-transport';
+import type { IRpcWebSocketTransport } from '@solana/rpc-types';
 
 import { getWebSocketTransportWithAutoping } from '../rpc-websocket-autopinger';
