+++ packages/rpc-transport/src/index.ts
@@ -5,5 +5,4 @@ export * from './json-rpc';
 export type { SolanaJsonRpcErrorCode } from './json-rpc-errors';
 export * from './json-rpc-subscription';
 export * from './transports/http/http-transport';
-export type { IRpcWebSocketTransport } from './transports/transport-types';
 export * from './transports/websocket/websocket-transport';
