--- packages/rpc-transport/src/transports/transport-types.ts
@@ -1,16 +1,0 @@
-import { RpcWebSocketConnection } from './websocket/websocket-connection';
-
-type RpcWebSocketTransportConfig = Readonly<{
-    payload: unknown;
-    signal: AbortSignal;
-}>;
-
-export interface IRpcWebSocketTransport {
-    (config: RpcWebSocketTransportConfig): Promise<
-        Readonly<
-            Omit<RpcWebSocketConnection, 'send'> & {
-                send_DO_NOT_USE_OR_YOU_WILL_BE_FIRED: RpcWebSocketConnection['send'];
-            }
-        >
-    >;
-}
